refactor to_exec_db
