+++ crates/vm/backends/mod.rs
@@ -375,13 +375,13 @@ impl Evm {
     pub fn to_exec_db(&mut self, block: &Block) -> Result<ExecutionDB, ExecutionDBError> {
         match self {
             Evm::REVM { state } => {
-                let sw = StoreWrapper {
+                let store_wrapper = StoreWrapper {
                     store: state.database().unwrap().clone(),
                     block_hash: block.header.parent_hash,
                 };
-                sw.to_exec_db_revm(block)
+                REVM::to_exec_db(&store_wrapper, block)
             }
-            Evm::LEVM { store_wrapper, .. } => store_wrapper.to_exec_db_levm(block),
+            Evm::LEVM { store_wrapper, .. } => LEVM::to_exec_db(store_wrapper, block),
         }
     }
 }
